Make cli-initialized apps start at build 1
Previously Quip would initialize apps at build 1, version 0.1.0 - this meant that the first visible version (since it has to be published AFTER the auto-created first version) would be v0.1.1, build 2. This was confusing to beta testers, so this commit (and a corresponding change in the Quip API) makes it so that the first version you see in the dev console after running init is version 0.1.0, build 1.

This also moves us to use the next tag for Quip dependencies in templates, which I'll update to a specific version once we GA.